+++ ChessPlayer/build.gradle
@@ -1,0 +1,45 @@
+plugins {
+    id 'java'
+}
+
+group 'org.example'
+version '1.0-SNAPSHOT'
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation files('../engine/stockfish/stockfish.exe')
+    implementation fileTree('../engine/stockfish/src') {include '*.h' }
+    implementation fileTree('../engine/stockfish/src') {include '*.cpp'}
+
+    implementation 'org.apache.commons:commons-text:1.9'
+    implementation 'commons-io:commons-io:2.11.0'
+    implementation 'org.apache.logging.log4j:log4j-core:2.18.0'
+
+    implementation 'com.google.guava:guava:31.1-jre'
+    implementation 'com.google.dagger:dagger:2.42'
+    annotationProcessor 'com.google.dagger:dagger-compiler:2.42'
+    implementation 'com.google.code.gson:gson:2.8.8'
+
+    implementation 'com.fasterxml.jackson.core:jackson-core:2.11.4'
+    implementation 'com.fasterxml.jackson.core:jackson-databind:2.11.4'
+
+    implementation 'com.amazonaws:aws-lambda-java-core:1.2.1'
+    implementation 'com.amazonaws:aws-lambda-java-events:3.11.0'
+    implementation 'com.amazonaws:aws-java-sdk-cloudwatch:1.12.+'
+    implementation 'com.amazonaws:aws-java-sdk-dynamodb:1.12.+'
+
+    implementation 'javax.measure:unit-api:2.1.3'
+
+    implementation 'org.mockito:mockito-core:4.+'
+    implementation "org.mockito:mockito-inline:4.+"
+
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
+    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
+}
+
+test {
+    useJUnitPlatform()
+}
