+++ ChessPlayer/src/main/java/com/nashss/se/chessplayerservice/dynamodb/models/Game.java
@@ -1,0 +1,41 @@
+package com.nashss.se.chessplayerservice.dynamodb.models;
+
+public class Game {
+
+    private String gameId;
+    private String moves;
+    private String whitePlayerId;
+    private String blackPlayerId;
+
+    public String getGameId() {
+        return gameId;
+    }
+
+    public void setGameId(String gameId) {
+        this.gameId = gameId;
+    }
+
+    public String getMoves() {
+        return moves;
+    }
+
+    public void setMoves(String moves) {
+        this.moves = moves;
+    }
+
+    public String getWhitePlayerId() {
+        return whitePlayerId;
+    }
+
+    public void setWhitePlayerId(String whitePlayerId) {
+        this.whitePlayerId = whitePlayerId;
+    }
+
+    public String getBlackPlayerId() {
+        return blackPlayerId;
+    }
+
+    public void setBlackPlayerId(String blackPlayerId) {
+        this.blackPlayerId = blackPlayerId;
+    }
+}
