+++ settings.gradle
@@ -1,2 +1,3 @@
 rootProject.name = 'nss-capstone-josh-taylor'
+include 'ChessPlayer'
