+++ template.yaml
@@ -1,0 +1,52 @@
+AWSTemplateFormatVersion: '2010-09-09'
+Transform: AWS::Serverless-2016-10-31
+Description: >
+  Chess Player Client
+
+Parameters:
+  S3Bucket:
+    Type: String
+    Description: Parameter to specify the S3 Bucket to use for deployment
+  FrontendDeployment:
+    Default: remote
+    Type: String
+    AllowedValues:
+      - local
+      - remote
+    ConstraintDescription: Must specify 'local' or 'remote' for FrontendDeployment.
+
+Conditions:
+  DeployCloudFront: !Equals
+    - !Ref FrontendDeployment
+    - remote
+
+Globals:
+  Function:
+    Timeout: 20
+
+  Api:
+    Cors:
+      AllowMethods: "'GET,POST,PUT,OPTIONS'"
+      AllowHeaders: "'content-type'"
+      AllowOrigin: "'*'"
+
+Resources:
+
+  GetNextMoveLambda:
+    Type: AWS::Serverless::Function
+    Properties:
+      CodeUri: ChessPlayer
+      Handler: com.nashss.se.chessplayerservice.lambda.GetNextMoveLambda::handleRequest
+      Runtime: java11
+      Architectures:
+        - x86_64
+      MemorySize: 512
+      Environment:
+        Variables:
+          JAVA_TOOL_OPTIONS: -XX:+TieredCompilation -XX:TieredStopAtLevel=1
+      Events:
+        ChessPlayerClient:
+          Type: Api
+          Properties:
+            Path: /move/{move}
+            Method: get
